refactor: replace FloatingWindow with ServiceOverlay for better service management
